+++ backend/app/api/v1/endpoints/likes.py
@@ -1,30 +1,33 @@
 from typing import Any, List
 
+from fastapi import APIRouter, HTTPException, Request
+from api.deps import auth_guard
 from crud import crud_likes
-
-from fastapi import APIRouter, HTTPException
 
 router = APIRouter()
 
 
 @router.get("/posts/{id}")
-def get_likes_by_post_id(id):
+def get_likes_by_post_id():
     try:
-        return crud_likes.CRUD_likes.get_by_comment_id(id)
+        # return crud_likes.CRUD_likes.get_by_comment_id(id)
+        return {}
     except HTTPException:
         pass 
 
 
 @router.get("/comments/{id}")
-def get_likes_by_comment_id(id):
+def get_likes_by_comment_id():
     try:
-        return crud_likes.CRUD_likes.get_by_comment_id(id)
+        # return crud_likes.CRUD_likes.get_by_comment_id(id)
+        return {}
     except HTTPException:
         pass 
 
 
 @router.post("/comments")
-def manage_like_on_comment():
+@auth_guard("user")
+def manage_like_on_comment(request: Request):
     try:
         return crud_likes.CRUD_likes.manage_on_comment
     except HTTPException:
@@ -32,7 +35,8 @@ def manage_like_on_comment():
 
 
 @router.post("/posts")
-def manage_like_on_post():
+@auth_guard("user")
+def manage_like_on_post(request: Request):
     try:
         return crud_likes.CRUD_likes.manage_on_post
     except HTTPException: