+++ backend/app/api/v1/endpoints/posts.py
@@ -1,5 +1,7 @@
 from typing import Any, List
-from fastapi import APIRouter, HTTPException
+
+from fastapi import APIRouter, HTTPException, Request
+from api.deps import auth_guard
 from crud import crud_posts
 
 
@@ -15,15 +17,17 @@ def get_posts():
 
 
 @router.get("/{id}")
-def get_post_by_id(id):
-    try:
-        return crud_posts.CRUD_posts.get_by_id(id)
-    except HTTPException:
-        pass 
+def get_post_by_id():
+	try:
+		# return crud_posts.CRUD_posts.get_by_id(id)
+		return {}
+	except HTTPException:
+		pass 
 
 
 @router.post("/")
-def create_post():
+@auth_guard("user")
+def create_post(request: Request):
     try:
         return crud_posts.CRUD_posts.create
     except HTTPException:
@@ -31,16 +35,20 @@ def create_post():
 
 
 @router.patch("/{id}")
-def modify_post():
-    try:
-        return crud_posts.CRUD_posts.modify(id)
-    except HTTPException:
-        pass 
+@auth_guard("user")
+def modify_post(request: Request):
+	try:
+		# return crud_posts.CRUD_posts.modify(id)
+		print(request.attach_user)
+	except HTTPException:
+		pass 
 
 
 @router.delete("/{id}")
-def delete_post(id):
-    try:
-        return crud_posts.CRUD_posts.delete(id)
-    except HTTPException:
-        pass 
+@auth_guard("user")
+def delete_post(request: Request):
+	try:
+		# return crud_posts.CRUD_posts.delete(id)
+		print(request.attach_user)
+	except HTTPException:
+		pass