+++ backend/app/api/v1/endpoints/reports.py
@@ -1,20 +1,26 @@
-from fastapi import APIRouter, HTTPException
-from crud import crud_reports
+from typing import Any, List
 
+from fastapi import APIRouter, HTTPException, Request
+from api.deps import auth_guard
+from crud import crud_reports
 
 router = APIRouter()
 
 
 @router.get("/")
-def get_report(id):
+@auth_guard("admin")
+def get_report(request: Request):
     try:
-        return crud_reports.CRUD_reports.get(id)
+        print(request)
+        print(request.attach_user)
+        return { "sucess": True }
     except HTTPException:
         pass 
 
 
 @router.post("/")
-def create_report():
+@auth_guard("user")
+def create_report(request: Request):
     try:
         return crud_reports.CRUD_reports.create
     except HTTPException:
@@ -22,8 +28,10 @@ def create_report():
 
 
 @router.delete("/{id}")
-def delete_report(id):
-    try:
-        return crud_reports.CRUD_reports.delete(id)
-    except HTTPException:
-        pass 
+@auth_guard("admin")
+def delete_report(request: Request):
+	try:
+		# return crud_reports.CRUD_reports.delete(id)
+		print(request.attach_user)
+	except HTTPException:
+		pass