+++ history.md
@@ -98,3 +98,10 @@ Updated schema
 Updated schema
 * added schema for likes
 * Fixed authors display
+
+## 0.2.17 (2022-10-13)
+Added auth guard decorator + integrations
+* specify user role
+* link to jwt decode and return user
+* generator yield for mongodb database
+* routes order fixed